1.7.4.2
**Bug Fixes**
Fixed type on [gamipress_points_types] shortcode slug used to pair
attributes.
**Improvements**
Improved user ID detection on user profile form included when other
plugins causes issues on this form to keep GamiPress utilities working.
Prevent to display all user earnings to visitors if current user is set
to "yes".
**Developer Notes**
Added filters to change the workflow on user earnings and logs shortcode
to display or not content to visitors when current user is set to "yes".